add docker file and reorganize src
