+++ scripts/build_linux.sh
@@ -29,7 +29,7 @@ echo " => Running PyInstaller to create executable..."
 pyinstaller --onefile \
     --hidden-import="zeroconf._utils.ipaddress" \
     --hidden-import="zeroconf._handlers.answers" \
-    --add-data="src/onthespot/gui/qtui/*.ui:onthespot/gui/qtui" \
+    --add-data="src/onthespot/qt/qtui/*.ui:onthespot/qt/qtui" \
     --add-data="src/onthespot/resources/icons/*.png:onthespot/resources/icons" \
     --add-data="src/onthespot/resources/translations/*.qm:onthespot/resources/translations" \
     $FFBIN \
