+++ scripts/build_mac.sh
@@ -24,7 +24,7 @@ unzip build/ffmpeg.zip -d build
 pyinstaller --windowed \
     --hidden-import="zeroconf._utils.ipaddress" \
     --hidden-import="zeroconf._handlers.answers" \
-    --add-data="src/onthespot/gui/qtui/*.ui:onthespot/gui/qtui" \
+    --add-data="src/onthespot/qt/qtui/*.ui:onthespot/qt/qtui" \
     --add-data="src/onthespot/resources/icons/*.png:onthespot/resources/icons" \
     --add-data="src/onthespot/resources/translations/*.qm:onthespot/resources/translations" \
     --add-binary="build/ffmpeg:onthespot/bin/ffmpeg" \
