+++ setup.cfg
@@ -21,7 +21,7 @@ install_requires = file:requirements.txt
 console_scripts =
     onthespot-cli = onthespot.cli:main
     onthespot-web = onthespot.web:main
-    onthespot-gui = onthespot.__init__:main
+    onthespot-gui = onthespot.gui:main
 
 [options.packages.find]
 where=src
