+++ src/onthespot/qt/minidialog.py
