+++ src/onthespot/qt/qtui/notice.ui
