+++ src/onthespot/qt/settings.py
