+++ src/onthespot/qt/thumb_listitem.py
