+++ src/portable.py
@@ -1,5 +1,5 @@
 #!/usr/bin/python3
-from onthespot import main
+from onthespot.gui import main
 
 if __name__ == '__main__':
     main()
